requirements.txt
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
gradio>=4.14.0
Pillow>=10.0.0
peft>=0.7.0
safetensors>=0.4.0